Initial progress towards Abrash
Big overhaul of the code to shift to Abrash's model, based on having
read https://ericlippert.com/2020/05/21/life-part-11/

This makes it a little bit more what I believe to be "Rusty", and will
enable me to ditch the constants in favour of some CLi arguments soon.
I'll tackle that once I get the Abrash approach actually working!

I've a bug in the process_frame logic somewhere that I need to unpick.
The simple blinker test case fails.

The rest of the tests in src/lib.rs also need ported to the new format,
as do the benchmarks.